Move test in the correct test module.
