+++ tests/test_dispatch.py
@@ -6,62 +6,6 @@ import pytest
 from multimethod import get_types, multidispatch, multimethod, signature, DispatchError
 
 
-def test_signature():
-    with pytest.raises(TypeError):
-        signature([list]) <= signature([None])
-
-
-# roshambo
-class rock: pass
-
-class paper: pass
-
-class scissors: pass
-
-
-@multimethod
-def roshambo(left, right):
-    return 'tie'
-
-
-@roshambo.register(scissors, rock)
-@roshambo.register(rock, scissors)
-def _(left, right):
-    return 'rock smashes scissors'
-
-
-@roshambo.register(paper, scissors)
-@roshambo.register(scissors, paper)
-def _(left, right):
-    return 'scissors cut paper'
-
-
-@roshambo.register(rock, paper)
-@roshambo.register(paper, rock)
-def _(left, right):
-    return 'paper covers rock'
-
-
-def test_roshambo():
-    assert roshambo.__name__ == 'roshambo'
-    r, p, s = rock(), paper(), scissors()
-    assert len(roshambo) == 7
-    assert roshambo(r, p) == 'paper covers rock'
-    assert roshambo(p, r) == 'paper covers rock'
-    assert roshambo(r, s) == 'rock smashes scissors'
-    assert roshambo(p, s) == 'scissors cut paper'
-    assert roshambo(r, r) == 'tie'
-    assert len(roshambo) == 8
-    del roshambo[()]
-    del roshambo[rock, paper]
-    assert len(roshambo) == 5
-    with pytest.raises(DispatchError, match="0 methods"):
-        roshambo(r, r)
-    r = roshambo.copy()
-    assert isinstance(r, multimethod)
-    assert r == roshambo
-
-
 def test_cls():
     class cls:
         method = multidispatch(lambda self, other: None)
