+++ tests/test_generics.py
@@ -1,0 +1,8 @@
+import pytest
+
+from multimethod import signature
+
+
+def test_signature():
+    with pytest.raises(TypeError):
+        signature([list]) <= signature([None])
